+++ app/(tabs)/index.tsx
@@ -1,78 +1,64 @@
-import { Image, StyleSheet, Platform } from "react-native";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { useState, useEffect, useRef } from "react";
+import * as Location from "expo-location";
+import { useCameraPermissions } from "expo-camera";
+import * as Notifications from "expo-notifications";
 
-import { HelloWave } from "@/components/HelloWave";
-import ParallaxScrollView from "@/components/ParallaxScrollView";
-import { ThemedText } from "@/components/ThemedText";
-import { ThemedView } from "@/components/ThemedView";
+import Map from "@/components/Map";
+import UserStore from "@/stores/UserStore";
+import { inject, observer } from "mobx-react";
 
-export default function HomeScreen() {
+interface Props {
+  _userStore?: UserStore;
+}
+
+function MapScreen({ _userStore }: Props) {
+  const userStore = _userStore!;
+
+  const [cameraPermission, requestCameraPermission] = useCameraPermissions();
+  const [locationPermissionStatus, setLocationPermissionStatus] = useState("");
+
+  const deviceId = useRef("");
+
+  Notifications.getDevicePushTokenAsync().then(({ data }) => {
+    console.log("deviceId:", data);
+    deviceId.current = data;
+    userStore.getCurrentUser().setDeviceId(data);
+  });
+
+  useEffect(() => {
+    async function getCurrentLocation() {
+      const { status } = await Location.requestForegroundPermissionsAsync();
+      return status;
+    }
+
+    getCurrentLocation().then(async (status) => {
+      setLocationPermissionStatus(status);
+    });
+  }, []);
+
+  useEffect(() => {
+    if (locationPermissionStatus === "granted") {
+      Location.getCurrentPositionAsync({});
+    }
+  }, [locationPermissionStatus]);
+
+  useEffect(() => {
+    if (cameraPermission && !cameraPermission.granted) {
+      requestCameraPermission();
+    }
+  }, [cameraPermission]);
+
+  const isDisplayMap = () => {
+    return locationPermissionStatus == "granted";
+  };
+
   return (
-    <ParallaxScrollView
-      headerBackgroundColor={{ light: "#A1CEDC", dark: "#1D3D47" }}
-      headerImage={
-        <Image
-          source={require("@/assets/images/partial-react-logo.png")}
-          style={styles.reactLogo}
-        />
-      }
-    >
-      <ThemedView style={styles.titleContainer}>
-        <ThemedText type="title">Welcome!</ThemedText>
-        <HelloWave />
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 1: Try it</ThemedText>
-        <ThemedText>
-          Edit{" "}
-          <ThemedText type="defaultSemiBold">app/(tabs)/index.tsx</ThemedText>{" "}
-          to see changes. Press{" "}
-          <ThemedText type="defaultSemiBold">
-            {Platform.select({
-              ios: "cmd + d",
-              android: "cmd + m",
-              web: "F12",
-            })}
-          </ThemedText>{" "}
-          to open developer tools.
-        </ThemedText>
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 2: Explore</ThemedText>
-        <ThemedText>
-          Tap the Explore tab to learn more about what's included in this
-          starter app.
-        </ThemedText>
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 3: Get a fresh start</ThemedText>
-        <ThemedText>
-          When you're ready, run{" "}
-          <ThemedText type="defaultSemiBold">npm run reset-project</ThemedText>{" "}
-          to get a fresh <ThemedText type="defaultSemiBold">app</ThemedText>{" "}
-          directory. This will move the current{" "}
-          <ThemedText type="defaultSemiBold">app</ThemedText> to{" "}
-          <ThemedText type="defaultSemiBold">app-example</ThemedText>.
-        </ThemedText>
-      </ThemedView>
-    </ParallaxScrollView>
+    <SafeAreaView>
+      {/* TODO: 設定画面ができた時に設定画面にマップを出力させたい */}
+      <Map mapVisible={isDisplayMap()} />
+    </SafeAreaView>
   );
 }
 
-const styles = StyleSheet.create({
-  titleContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 8,
-  },
-  stepContainer: {
-    gap: 8,
-    marginBottom: 8,
-  },
-  reactLogo: {
-    height: 178,
-    width: 290,
-    bottom: 0,
-    left: 0,
-    position: "absolute",
-  },
-});
+export default inject("_userStore")(observer(MapScreen));
